requirements.txt
Flask
gunicorn
mysql-connector-python
PyMySQL
cryptography